Fix a bug where folder's existence was not checked for before removing

+++ OmnifactoryMigrate.py
@@ -13,16 +13,19 @@ new_path=input("Enter the path to the new version's minecraft folder.\n")
 
 #This will take the important folders/files recommended on the Omnifactory GitHub and copy them to the new instance folder.
 def copyFiles(old_path,new_path):
-    shutil.rmtree(os.path.join(new_path,'saves'))
+    if os.path.exists(os.path.join(new_path,'saves')):
+        shutil.rmtree(os.path.join(new_path,'saves'))
     shutil.copytree(os.path.join(old_path,'saves'), os.path.join(new_path,'saves'))
     
-    shutil.rmtree(os.path.join(new_path,'resourcepacks'))
+    if os.path.exists(os.path.join(new_path,'resourcepacks')):
+        shutil.rmtree(os.path.join(new_path,'resourcepacks'))
     shutil.copytree(os.path.join(old_path,'resourcepacks'), os.path.join(new_path,'resourcepacks'))
     
     os.remove(os.path.join(new_path,'options.txt'))
     shutil.copyfile(os.path.join(old_path,'options.txt'), os.path.join(new_path,'options.txt'))
     
-    shutil.rmtree(os.path.join(new_path,'journeymap'))
+    if os.path.exists(os.path.join(new_path,'journeymap')):
+        shutil.rmtree(os.path.join(new_path,'journeymap'))
     shutil.copytree(os.path.join(old_path,'journeymap'), os.path.join(new_path,'journeymap'))
     
     os.remove(os.path.join(new_path,'config','jei','bookmarks.ini'))
